Temporary commit...
There are errors in Moritz.Symbols, that need correcting first.
Moritz.Composer exists but has been made unavailable, and needs work.
Form1 needs a line that constructs the score (when the constructor arguments have been finalized).

+++ Moritz.Symbols/Metrics/Metrics_Text.cs
@@ -91,7 +91,7 @@ namespace Moritz.Symbols
 			Size maxSize = new Size(int.MaxValue, int.MaxValue);
 			TextFormatFlags flags = TextFormatFlags.NoPadding;
 			Size sizeOfString;
-			using(Font sysFont = new Font(fontFace, fontHeight))
+			using(Font sysFont = new Font(fontFace, (float)fontHeight))
 			{
 				sizeOfString = TextRenderer.MeasureText(g, text, sysFont, maxSize, flags);
 			}
